% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{gap}
\alias{gap}
\title{Fictitious dataset with importance and satisfaction results across five different
offices.}
\format{a data frame with 68 ovservations of 11 variables.}
\description{
This data set is used in the \code{GapAnalysis} demo and is used to demonstrate
how the \code{likert} package handles a gap analysis.
}
\keyword{datasets}

