% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_wrap_mod.r
\name{label_wrap_mod}
\alias{label_wrap_mod}
\title{Wrap label text.}
\usage{
label_wrap_mod(value, width = 25)
}
\arguments{
\item{value}{vector (converted using \code{\link{as.character}}) to be wrapped.}

\item{width}{the maximum width of each line in characters.

Adapted from https://github.com/hadley/ggplot2/wiki/labeller}
}
\description{
Wrap label text.
}

