\name{likert.density.plot}
\alias{likert.density.plot}
\title{Creates a density plot for likert items.}
\usage{
  likert.density.plot(likert, ordered = FALSE, ...)
}
\arguments{
  \item{likert}{object of type likert.}

  \item{ordered}{reorder items from high to low.}

  \item{...}{currently unused.}
}
\description{
  This funciton will create a visualization that treats the
  likert items as a continuous variable.
}
\seealso{
  plot.likert
}

