% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likert.options.R
\name{likert.options}
\alias{likert.options}
\title{Builds an object with options for plotting `likert` results.}
\usage{
likert.options(
  low.color = "#D8B365",
  high.color = "#5AB4AC",
  neutral.color = "grey90",
  neutral.color.ramp = "white",
  colors = NULL,
  plot.percent.low = TRUE,
  plot.percent.high = TRUE,
  plot.percent.neutral = TRUE,
  plot.percents = FALSE,
  text.size = 3,
  text.color = "black",
  text.color.manual.pos = NULL,
  text.color.manual.neg = NULL,
  text.color.manual.neutral = NULL,
  centered = TRUE,
  include.center = TRUE,
  ordered = TRUE,
  wrap = 50,
  wrap.grouping = 50,
  legend = "Response",
  ylabel = "Percentage",
  legend.position = c("bottom", "top", "left", "right", "none"),
  panel.arrange = "v",
  panel.strip.color = "#F0F0F0",
  digits = 2,
  drop0trailing = FALSE,
  zero.print = TRUE,
  ...
)
}
\arguments{
\item{low.color}{color for low values.}

\item{high.color}{color for high values.}

\item{neutral.color}{color for middle values (if odd number of levels).}

\item{neutral.color.ramp}{second color used when calling [colorRamp()].
with \code{low.color} and \code{high.color} to define the color palettes.}

\item{colors}{vector specifying the colors to use. This must be equal to 
the number of likert levels.}

\item{plot.percent.low}{whether to plot low percentages.}

\item{plot.percent.high}{whether to plot high percentages.}

\item{plot.percent.neutral}{whether to plot netural percentages.}

\item{plot.percents}{whether to label each category/bar.}

\item{text.size}{size of text attributes.}

\item{text.color}{color of text attributes.}

\item{text.color.manual.pos}{individual color specification of positive text attributes.}

\item{text.color.manual.neg}{individual color specification of negative text attributes.}

\item{text.color.manual.neutral}{individual color specification of neutral text attributes.}

\item{centered}{if true, the bar plot will be centered around zero such that
the lower half of levels will be negative.}

\item{include.center}{if TRUE, include the center level in the plot otherwise
the center will be excluded.}

\item{ordered}{reorder items from high to low.}

\item{wrap}{width to wrap label text for item labels}

\item{wrap.grouping}{width to wrap label text for group labels.}

\item{legend}{title for the legend.}

\item{ylabel}{label for the y-axis}

\item{legend.position}{the position for the legend ("left", "right", "bottom",
"top", or two-element numeric vector).}

\item{panel.arrange}{how panels for grouped likert items should be arrange.
Possible values are \code{v} (vertical, the default), \code{h}
(horizontal), and \code{NULL} (auto fill horizontal and vertical)}

\item{panel.strip.color}{the background color for panel labels.}

\item{digits}{the number of significant digits to print.}

\item{drop0trailing}{logical, indicating if trailing zeros, i.e., "0" after the decimal mark, should be removed}

\item{zero.print}{logical, character string or NULL specifying if and how zeros should be formatted specially.}

\item{...}{included for future expansion.}
}
\description{
This function specifies all the plotting options for Likert plots.
}
