% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likert_anchors.R
\docType{data}
\name{likert_anchors}
\alias{likert_anchors}
\title{Sample Likert scale response anchors.}
\format{
list of character vectors corresponding to the anchors. The names of the list provide
        a description of the anchor types.
}
\description{
This is a list of sample response anchors for Likert scales. This was adapted from Vagias (2006).
}
\references{
Vagias, Wade M. (2006). Likert-type scale response anchors. Clemson International 
Institute for Tourism & Research Development, Department of Parks, Recreation and Tourism 
Management. Clemson University.
}
