% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/likert-package.R
\docType{data}
\name{MathAnxietyGender}
\alias{MathAnxietyGender}
\title{Pre-summarized results from an administration of the Math Anxiety Scale Survey
grouped by gender.}
\format{data frame with 28 rows and 7 columns.}
\description{
A data frame of presummarized results of the Math Anxiety Scale Survey administered
to 20 students in a statistics course grouped by gender.
}
\references{
Bai, H., Wang, L., Pan, W., & Frey, M. (2009). Measuring mathematics
anxiety: Psychometric analysis of a bidimensional affective scale. Journal of
Instructional Psychology, 36 (3), 185- 193.
}
\keyword{datasets}

