\docType{data}
\name{pisa}
\alias{pisa}
\title{North American (i.e. Canada, Mexico, and United States) results from the 2009
Programme of International Student Assessment (PISA)
as provided by the Organization for Economic Co-operation and Development (OECD).
See \url{http://www.pisa.oecd.org/} for more information including the code book.}
\format{a data frame 66,690 ovservations of 437 variables from North America.}
\source{
  Organization for Economic Co-operation and Development
}
\description{
  North American (i.e. Canada, Mexico, and United States)
  results from the 2009 Programme of International Student
  Assessment (PISA) as provided by the Organization for
  Economic Co-operation and Development (OECD). See
  \url{http://www.pisa.oecd.org/} for more information
  including the code book.
}
\keyword{datasets}

