% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likert-package.R
\docType{data}
\name{pisaitems}
\alias{pisaitems}
\title{Programme of International Student Assessment}
\format{
a data frame 66,690 observations of 81 variables from North America.
}
\source{
Organization for Economic Co-operation and Development
}
\usage{
pisaitems
}
\description{
North American (i.e. Canada, Mexico, and United States) results from the 2009
Programme of International Student Assessment (PISA)
as provided by the Organization for Economic Co-operation and Development (OECD).
See \url{https://www.oecd.org/pisa/} for more information including the code book.
}
\keyword{datasets}
