% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.likert.bar.r
\name{print.likert.bar.plot}
\alias{print.likert.bar.plot}
\title{Print method for \code{\link{likert.bar.plot}}. The primary purpose is to 
suppress the "Stacking not well defined when ymin != 0" warning printed
by \code{ggplot2} for bar plots that have negative bars (i.e. the centered
plots).}
\usage{
\method{print}{likert.bar.plot}(x, ...)
}
\arguments{
\item{x}{a plot from \code{\link{likert.bar.plot}}.}

\item{...}{other parameters passed to ggplot2.}
}
\description{
Print method for \code{\link{likert.bar.plot}}. The primary purpose is to 
suppress the "Stacking not well defined when ymin != 0" warning printed
by \code{ggplot2} for bar plots that have negative bars (i.e. the centered
plots).
}

