\name{print.likert.heat.plot}
\alias{print.likert.heat.plot}
\title{Print method for \code{\link{plot.likert.heat}}.}
\usage{
  print.likert.heat.plot(p, ...)
}
\arguments{
  \item{p}{a plot from \code{\link{plot.likert.heat}}.}

  \item{...}{other parameters passed to ggplot2.}
}
\description{
  Print method for \code{\link{plot.likert.heat}}.
}

