% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtable.likert.r
\name{print.xlikert}
\alias{print.xlikert}
\title{Prints the results of `xtable.likert`.}
\usage{
\method{print}{xlikert}(x, tabular.environment = "longtable", floating = FALSE, ...)
}
\arguments{
\item{x}{results of [xtable.likert()].}

\item{tabular.environment}{see [xtable::print.xtable()].}

\item{floating}{see [print.xtable()].}

\item{...}{other parameters passed to [xtable::print.xtable()]}
}
\description{
Print method for [xtable.likert()].
}
