% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likert-package.R
\docType{data}
\name{sasr}
\alias{sasr}
\title{Results from the Survey of Academic Self-Regulation (SASR).}
\format{a data frame with 860 ovservations of 63 variables.}
\description{
The Survey of Academic Self-Regulation (SASR) is comprised of six factors:
self-regulation, intrinsic motivation, extrinsic motivation, self-efficacy,
metacognition, and personal relevance and control.
}
\references{
Dugan, R., & Andrade, H. (2011). Exploring the construct validity 
of academic self-regulation using a new self-report questionnaire. The 
International Journal of Educational and Psychological Assessment, 7(1).
}
\keyword{datasets}

