% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyLikert.R
\name{shinyLikert}
\alias{shinyLikert}
\title{Shiny App for the likert package.}
\usage{
shinyLikert()
}
\description{
This will start a shiny app included with the package to show many of the features
available in the likert package.
}
\references{
http://rstudio.com/shiny
}

