% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.likert.gap.R
\name{summary.likert.gap}
\alias{summary.likert.gap}
\title{Prints summary table of a Likert analysis.}
\usage{
\method{summary}{likert.gap}(object, ...)
}
\arguments{
\item{object}{the likert class to summarize.}

\item{...}{parameters passed to \code{\link{summary.likert}}}
}
\value{
a list with two data frames with summarized data for satisfaction
        and importance results separately.
}
\description{
The \code{summary} function returns a data frame that provides additional 
information. It contains 'Item' and 'Group' columns similiar to the results data
frame as well as a column 'low' corresponding to the sum of levels below
neutral, a column 'high' corresponding to the sum of levels above
neutral, and columns 'mean' and 'sd' corresponding to the mean and
standard deviation, respectively, of the results. The numeric values
are determined by as.numeric which will use the values of the factors.
}

