\name{xtable.likert}
\alias{xtable.likert}
\title{Prints a LaTeX table of the likert items.}
\usage{
  xtable.likert(x, caption = NULL, label = NULL,
    align = NULL, digits = NULL, display = NULL, ...)
}
\arguments{
  \item{x}{likert class object.}

  \item{caption}{the table caption.}

  \item{label}{the table label.}

  \item{align}{column alignments.}

  \item{digits}{number of digits to use for numeric
  columns.}

  \item{display}{column formats.}

  \item{...}{other parameters passed to \link{xtable}.}
}
\description{
  Prints a LaTeX table of the likert items.
}
\seealso{
  \link{xtable}, \link{print.xtable}
}

