% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/align.R
\name{zeroGrob}
\alias{zeroGrob}
\title{The zero grob draws nothing and has zero size.}
\usage{
zeroGrob()
}
\description{
The zero grob draws nothing and has zero size.
}
\author{
hadley
}
\references{
https://github.com/hadley/ggplot2/blob/master/R/grob-null.r
}
\keyword{internal}

