\name{abs_formatter}
\alias{abs_formatter}
\title{Absolute value formatter for \code{\link{continuous_scale}}.}
\usage{
  abs_formatter(x)
}
\arguments{
  \item{x}{value to be reformatted.}
}
\description{
  This will print the absolute value for labeling on axis.
  Usefull for stacked bar plots where negative values are
  not negative percentages but represent negative groups.
}

