% Generated by roxygen2 (4.0.2): do not edit by hand
\name{abs_formatter}
\alias{abs_formatter}
\title{Absolute value formatter for continuous_scale}
\usage{
abs_formatter(x)
}
\arguments{
\item{x}{value to be reformatted.}
}
\value{
the absolute value of x.
}
\description{
This will print the absolute value for labeling on axis. Usefull for stacked
bar plots where negative values are not negative percentages but represent
negative groups.
}

