% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/align.R
\name{align.plots}
\alias{align.plots}
\title{Adapted from ggExtra package which is no longer available. This is related to
an experimental mlpsa plot that will combine the circular plot along with
the two individual distributions.}
\usage{
\method{align}{plots}(gl, ...)
}
\arguments{
\item{gl}{grid.layout}

\item{...}{graphic elements to combine.}
}
\description{
Adapted from ggExtra package which is no longer available. This is related to
an experimental mlpsa plot that will combine the circular plot along with
the two individual distributions.
}
\references{
http://groups.google.com/group/ggplot2/browse_thread/thread/1b859d6b4b441c90
             http://ggextra.googlecode.com/svn/trunk/R/align.r
}

