% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likert-package.R
\docType{package}
\name{likert-package}
\alias{likert-package}
\title{Likert Analysis and Visualization}
\description{
Likert Analysis and Visualization
}
\seealso{
Useful links:
\itemize{
  \item \url{http://jason.bryer.org/likert}
  \item \url{http://github.com/jbryer/likert}
  \item Report bugs at \url{https://github.com/jbryer/likert/issues}
}

}
\author{
\email{jason@bryer.org}
}
\keyword{institutional}
\keyword{likert}
\keyword{package}
\keyword{research}
