% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likert-package.R
\docType{package}
\name{likert-package}
\alias{likert-package}
\title{Likert Analysis and Visualization}
\description{
Likert Analysis and Visualization
}
\author{
\email{jason@bryer.org}
}
\keyword{institutional}
\keyword{likert}
\keyword{package}
\keyword{research}

