% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.likert.bar.r
\name{likert.bar.plot}
\alias{likert.bar.plot}
\title{Bar Plot for Likert Items.}
\usage{
likert.bar.plot(l, group.order, center = (l$nlevels - 1)/2 + 1, ...)
}
\arguments{
\item{l}{results of \code{\link{likert}}.}

\item{group.order}{the order in which groups (for grouped items) or items
(for non-grouped items) should be plotted.}

\item{center}{specifies which level should be treated as the center. For example,
\code{center = 3} would use the third level as the center whereas
\code{center = 3.5} would indicate no specific level is the center but
<= 3 are low levels and >= 4 are high levels (i.e. used for forced choice 
items or those without a neutral option). This also influences where the
color breaks from low to high.}

\item{...}{passed to \code{\link{likert.options}}}

\item{likert}{object of type likert.}
}
\description{
Bar plot for the results of \code{\link{likert}}.
}
\seealso{
plot.likert

likert.heat.plot

likert.bar.plot

likert.density.plot
}

