\name{likert.density.plot}
\alias{likert.density.plot}
\title{Creates a density plot for likert items.}
\usage{
  likert.density.plot(likert, facet = TRUE, ...)
}
\arguments{
  \item{likert}{object of type likert.}

  \item{facet}{for non-grouped items, should each density
  distribution be plotted in a separate facet.}

  \item{...}{currently unused.}
}
\description{
  This funciton will create a visualization that treats the
  likert items as a continuous variable.
}
\seealso{
  plot.likert
}

