\name{likert.heat.plot}
\alias{likert.heat.plot}
\title{Internal method.}
\usage{
  likert.heat.plot(likert, low.color = "white",
    high.color = "blue", text.color = "black",
    text.size = 4, wrap = 50, ...)
}
\arguments{
  \item{likert}{object of type likert.}

  \item{low.color}{color for low values.}

  \item{high.color}{color for high values.}

  \item{text.size}{size of text attributes.}

  \item{text.color}{color of text attributes.}

  \item{wrap}{width to wrap label text for non-grouped
  likert objects.}

  \item{...}{currently unused.}
}
\description{
  Internal method.
}
\seealso{
  plot.likert

  likert.bar.plot
}

