\name{plot.likert.bar}
\alias{plot.likert.bar}
\title{Internal method.}
\usage{
  plot.likert.bar(likert, low.color = "blue",
    high.color = "red", neutral.color = "white",
    text.size = 3, text.color = "black", centered = FALSE,
    ...)
}
\arguments{
  \item{centered}{if true, the bar plot will be centered
  around zero such that the lower half of levels will be
  negative.}
}
\description{
  Internal method.
}
\seealso{
  plot.likert
}

