% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print.likert.gap.R
\name{print.likert.gap}
\alias{print.likert.gap}
\title{Prints results table.}
\usage{
\method{print}{likert.gap}(x, ...)
}
\arguments{
\item{x}{the likert class to print.}

\item{...}{parameters passed to \code{\link{print.data.frame}}.}
}
\description{
Prints results table.
}

