% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.likert.heat.r
\name{print.likert.heat.plot}
\alias{print.likert.heat.plot}
\title{Print method for [likert.heat.plot()].}
\usage{
\method{print}{likert.heat.plot}(p, ...)
}
\arguments{
\item{p}{a plot from [likert.heat.plot().}

\item{...}{other parameters passed to ggplot2.}
}
\description{
Print method for [likert.heat.plot()].
}
