% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.likert.r
\name{print.likert}
\alias{print.likert}
\title{Prints results table.}
\usage{
\method{print}{likert}(x, ...)
}
\arguments{
\item{x}{the likert class to print.}

\item{...}{parameters passed to \code{\link{print.data.frame}}.}
}
\description{
Prints results table.
}

